Enabling Assert (why was it disabled?)
git-svn-id: https://epos.lisha.ufsc.br/svn/epos2@4095 86035c5f-d46c-4136-a2d5-fe63f1d4a782

+++ branches/merge/tstp/include/utility/malloc.h
@@ -64,7 +64,7 @@ private:
 
 public:
     static void * alloc(unsigned int bytes, const EPOS::Page_Coloring_Allocator & allocator) {
-        assert(allocator <= COLORS);
+        assert((unsigned int)allocator <= COLORS);
         return _heap[allocator]->alloc(bytes);
     }
